Documentation improvements
* Rewrite README

- Focus on common-path, remove complex subjects that can be left to
  Haddocks and unit tests
- Replace generic models with a real domain
- Remove some comments, rewrite others

* Re-org modules
* Re-document extracted modules
* Re-order and re-document Graphula exports
* Fix weeder on CI
