+++ .github/workflows/ci.yml
@@ -11,4 +11,6 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: freckle/stack-cache-action@v1
-      - uses: freckle/stack-action@v1.1
+      - uses: freckle/stack-action@main
+        with:
+          weeder: false
