+++ graphula.cabal
@@ -1,10 +1,10 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.33.0.
+-- This file has been generated from package.yaml by hpack version 0.34.4.
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: cc5a03b69c515cc68baafe868c82fac1dc11e25d65e1349919b9650ce0e37ae2
+-- hash: d3fb852f80fa19f57cf24d5c4dc57c2444154e0b5f68e48e511299484a08ce1f
 
 name:           graphula
 version:        2.0.0.5
@@ -29,8 +29,14 @@ library
   exposed-modules:
       Graphula
       Graphula.Arbitrary
-      Graphula.Internal
+      Graphula.Class
+      Graphula.Dependencies
+      Graphula.Dependencies.Generic
+      Graphula.Idempotent
       Graphula.Key
+      Graphula.Logged
+      Graphula.NoConstraint
+      Graphula.Node
   other-modules:
       Paths_graphula
   hs-source-dirs:
@@ -60,7 +66,6 @@ test-suite readme
   type: exitcode-stdio-1.0
   main-is: README.lhs
   other-modules:
-      Graphula.UUIDKey
       Paths_graphula
   hs-source-dirs:
       test
@@ -71,6 +76,7 @@ test-suite readme
     , base <5
     , bytestring
     , containers
+    , generic-arbitrary
     , graphula
     , hspec
     , http-api-data
@@ -84,7 +90,6 @@ test-suite readme
     , text
     , transformers
     , unliftio-core
-    , uuid
   if impl(ghc >= 8.10)
     ghc-options: -Wno-missing-safe-haskell-mode -Wno-prepositive-qualified-module
   default-language: Haskell2010
