+++ package.yaml
@@ -56,6 +56,7 @@ tests:
       - aeson
       - bytestring
       - containers
+      - generic-arbitrary
       - graphula
       - hspec
       - http-api-data
@@ -69,4 +70,3 @@ tests:
       - text
       - transformers
       - unliftio-core
-      - uuid
